requirements.txt
Pillow==10.2.0
moviepy==1.0.3 